% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups_count.R
\name{dups_count}
\alias{dups_count}
\title{Count the total number of duplicate observations in a dataset}
\usage{
dups_count(data, ...)
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{id variables}
}
\value{

}
\description{
Count the total number of duplicate observations in a dataset
}
