% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups_drop.R
\name{dups_drop}
\alias{dups_drop}
\title{Drop duplicates}
\usage{
dups_drop(data, ..., warn = TRUE)
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{id variables}

\item{warn}{logical, when TRUE (default), the function will warn you about how many observations are being dropped}
}
\value{

}
\description{
Drop duplicates
}
