% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups_filter.R
\name{dups_filter}
\alias{dups_filter}
\title{Filter to look at only non-unique rows based on id variables}
\usage{
dups_filter(data, ...)
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{id variables}
}
\value{

}
\description{
Filter a dataset to only look at rows that are duplicated, useful for diagnosing why there may be duplicates
in the data and therefore how to remove them
}
