% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups_report.R
\name{dups_report}
\alias{dups_report}
\title{Print a duplicates report}
\usage{
dups_report(data, ..., print_only = TRUE)
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{id variables}

\item{print_only}{logical. Does the report get printed and then return the original dataframe (default), or
does it return the tibble with the duplicate details?}
}
\value{

}
\description{
Prints a reports of all the duplicates in a dataframe based on the uniquely identifying
variable(s)
}
