% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups_warn.R
\name{dups_stop}
\alias{dups_stop}
\title{Stop if there are duplicates in a dataset}
\usage{
dups_stop(data, ...)
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{id variables}
}
\value{

}
\description{
For use within functions, stops operations and throws an error if there are any duplicates in the data
}
