% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups_tag.R
\name{dups_tag}
\alias{dups_tag}
\title{Tag all observations that are duplicates}
\usage{
dups_tag(data, ..., var = "dups_tag")
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{id variables}

\item{var}{name of the new logical variable that will be created (that identifies the duplicates in the dataset)}
}
\value{

}
\description{
Returns the dataframe with a new logical variable that denotes whether the observation
is a duplicate observation or not
}
