% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups_view.R
\name{dups_view}
\alias{dups_view}
\title{View duplicate rows in the viewing portal}
\usage{
dups_view(data, ..., n_viewing = 200)
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{id variables}

\item{n_viewing}{randomly select n_viewing observations to view (useful for large datasets)}
}
\value{

}
\description{
View duplicate rows in the viewing portal
}
