% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups_warn.R
\name{dups_warn}
\alias{dups_warn}
\title{Warn if there are duplicates in a dataset}
\usage{
dups_warn(data, ...)
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{id variables}
}
\value{

}
\description{
For use within functions, warns the user if there are any duplicates in the data
}
