% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_n.R
\name{sample_n_groups}
\alias{sample_n_groups}
\title{Sample n groups in a grouped dataset}
\usage{
sample_n_groups(data, n)
}
\arguments{
\item{n}{numeric, how many groups to sample}
}
\value{

}
\description{
Sample n groups in a grouped dataset
}
