% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_n.R
\name{view_filter}
\alias{view_filter}
\title{View a randomly selected sample of observations after applying a filter}
\usage{
view_filter(data, ..., n = 200, by_group = TRUE)
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{arguments to pass to filter}

\item{n}{number of}

\item{by_group}{If TRUE, then view n randomly selected groups, if false then view n randomly selected rows}
}
\value{

}
\description{
View a randomly selected sample of observations after applying a filter
}
