% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_n.R
\name{view_n}
\alias{view_n}
\title{Randomly view N rows}
\usage{
view_n(data, n = 200, by_group = TRUE)
}
\arguments{
\item{data}{data.frame or tibble}

\item{n}{How many rows (or groups) to view}

\item{by_group}{If TRUE, then view n randomly selected groups, if false then view n randomly selected rows}
}
\value{

}
\description{
Randomly view N rows
}
