% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_n.R
\name{view_select}
\alias{view_select}
\title{View a randomly selected sample of observations after selecting only some variables to view}
\usage{
view_select(data, ..., n = 200, by_group = TRUE)
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{arguments to pass to select}

\item{n}{number of observations to view}

\item{by_group}{If TRUE, then view n randomly selected groups, if false then view n randomly selected rows}
}
\value{

}
\description{
View a randomly selected sample of observations after selecting only some variables to view
}
